# openapi.yaml
openapi: 3.0.3
info:
  title: DopUsdExchangeRate API
  version: 1.0.0
  description: API to fetch and calculate USD/DOP exchange rates from Dominican banks
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /sync:
    post:
      summary: Sync rates from all supported banks
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
  /rates:
    get:
      summary: Get all latest rates, ordered by cheapest buy rate
      responses:
        '200':
          description: List of rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRate'
  /rates/{bank}:
    get:
      summary: Get latest rates for a specific bank
      parameters:
        - name: bank
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of rates for the bank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRate'
        '404':
          description: Bank not supported
  /banks:
    get:
      summary: List supported banks
      responses:
        '200':
          description: List of banks
          content:
            application/json:
              schema:
                type: object
                properties:
                  banks:
                    type: array
                    items:
                      type: string
  /banks/{bank}/status:
    post:
      summary: Enable or disable a bank
      parameters:
        - name: bank
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disabled:
                  type: boolean
      responses:
        '200':
          description: Bank status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  bank:
                    type: string
                  disabled:
                    type: boolean
        '404':
          description: Bank not found
  /buy:
    get:
      summary: Convert DOP to USD for each bank using the bank's "venta" (sell) rate, and show calculated average
      description: For each bank, the buy_rate is the rate at which you buy USD from the bank (bank sells USD to you). This is taken from the bank's "venta" (sell) rate.
      parameters:
        - name: dop
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Conversion results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        bank:
                          type: string
                        usd:
                          type: number
                        buy_rate:
                          type: number
                  calculated:
                    type: object
                    properties:
                      avg_buy_rate:
                        type: number
                      usd:
                        type: number
  /sell:
    get:
      summary: Convert USD to DOP for each bank using the bank's "compra" (buy) rate, and show calculated average
      description: For each bank, the sell_rate is the rate at which you sell USD to the bank (bank buys USD from you). This is taken from the bank's "compra" (buy) rate.
      parameters:
        - name: usd
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Conversion results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        bank:
                          type: string
                        dop:
                          type: number
                        sell_rate:
                          type: number
                  calculated:
                    type: object
                    properties:
                      avg_sell_rate:
                        type: number
                      dop:
                        type: number
components:
  schemas:
    ExchangeRate:
      type: object
      properties:
        bank:
          type: string
        buy_rate:
          type: number
        sell_rate:
          type: number
        sync_date:
          type: string
          format: date-time